pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Binance Trading Bot with Technical Analysis"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pandas-ta>=0.4.71b0",
    "python-binance>=1.0.32",
]
